requirements.txt
supabase>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
yfinance>=0.2.30
schedule>=1.2.0
pytest>=7.0.0
